requirements-training.txt
# Training-specific dependencies for YOLOv8 finetuning

# Core ML libraries
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Experiment tracking
tensorboard>=2.13.0
mlflow>=2.8.0

# Jupyter for experimentation
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0

# YOLO utilities (already in requirements.txt)
# ultralytics - included in base requirements

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# YAML parsing
PyYAML>=6.0

# Progress bars
tqdm>=4.65.0

# Neural MMO (for research experiments)
nmmo>=2.0.0
pettingzoo>=1.24.0
gymnasium>=0.29.0

# Reinforcement learning (optional)
stable-baselines3>=2.1.0
vec-noise>=1.0.0

# Utilities
requests>=2.31.0
wandb>=0.15.0  # Alternative to MLflow

# Data annotation tools integration
label-studio-sdk>=0.0.30
